requirements.txt
CustomTkinter
CTkTable
joblib
numpy
pandas
Pillow
psycopg2
pubchempy
rdkit-pypi
scikit-learn